requirements.txt
pyyaml
requests
Flask
Flask-WTF[email]
Flask-Mail
Frozen-Flask
